mise a jour document

+++ Projet_PAF/src/main/java/com/inti/formation/IMetier/IDocumentMetier.java
@@ -2,6 +2,6 @@ package com.inti.formation.IMetier;
 
 import com.inti.formation.Model.Document;
 
-public interface IDocumentMetier extends IGenericMetier<Document, Integer> {
+public interface IDocumentMetier extends IGenericMetier<Document, Long> {
 
 }
