+++ Projet_PAF/src/main/java/com/inti/formation/IMetier/IUtilisateurMetier.java
@@ -2,6 +2,6 @@ package com.inti.formation.IMetier;
 
 import com.inti.formation.Model.Utilisateur;
 
-public interface IUtilisateurMetier extends  IGenericMetier<Utilisateur, Integer> {
+public interface IUtilisateurMetier extends  IGenericMetier<Utilisateur, Long> {
 
 }
