+++ Projet_PAF/src/main/java/com/inti/formation/repository/IDocumentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.inti.formation.Model.Document;
 
 @Repository("docRepo")
-public interface IDocumentRepository extends JpaRepository<Document, Integer> {
+public interface IDocumentRepository extends JpaRepository<Document, Long> {
 
 }
